+++ src/core/index.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  *
  */
-
 export * from './barrel/index.js';
 export * from './io/index.js';
 export * from './parser/index.js';